Consolidate controllers
